Combining container tests.
